logic for projecting via UMAP or PCA moved to separate dedicated Task classes.
