+++ src/main/java/edu/jhuapl/trinity/javafx/javafx3d/tasks/ProjectUmapFeaturesTask.java
@@ -1,0 +1,165 @@
+package edu.jhuapl.trinity.javafx.javafx3d.tasks;
+
+/*-
+ * #%L
+ * trinity
+ * %%
+ * Copyright (C) 2021 - 2023 The Johns Hopkins University Applied Physics Laboratory LLC
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import edu.jhuapl.trinity.data.messages.FeatureCollection;
+import edu.jhuapl.trinity.data.messages.FeatureVector;
+import edu.jhuapl.trinity.javafx.components.radial.ProgressStatus;
+import edu.jhuapl.trinity.javafx.events.ApplicationEvent;
+import edu.jhuapl.trinity.utils.AnalysisUtils;
+import edu.jhuapl.trinity.utils.ResourceUtils;
+import edu.jhuapl.trinity.utils.umap.Umap;
+import javafx.application.Platform;
+import javafx.concurrent.Task;
+import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.DialogPane;
+import javafx.scene.layout.Background;
+import javafx.scene.paint.Color;
+import javafx.stage.StageStyle;
+
+/**
+ * @author Sean Phillips
+ */
+public class ProjectUmapFeaturesTask extends Task {
+    Scene scene;
+    FeatureCollection originalFC;
+    Umap umap;
+    private boolean cancelledByUser = false;
+    private double projectionScalar = 100.0; //used for sizing values to 3D scene later
+
+    public ProjectUmapFeaturesTask(Scene scene, FeatureCollection originalFC, Umap umap) {
+        this.scene = scene;
+        this.originalFC = originalFC;
+        this.umap = umap;
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION,
+            "", ButtonType.YES, ButtonType.NO);
+        alert.setHeaderText("Watch a TV while you wait?");
+        alert.setGraphic(ResourceUtils.loadIcon("retrowave-tv", 100));
+        alert.initStyle(StageStyle.TRANSPARENT);
+        DialogPane dialogPane = alert.getDialogPane();
+        dialogPane.setBackground(Background.EMPTY);
+        dialogPane.getScene().setFill(Color.TRANSPARENT);
+        String DIALOGCSS = this.getClass().getResource("/edu/jhuapl/trinity/css/dialogstyles.css").toExternalForm();
+        dialogPane.getStylesheets().add(DIALOGCSS);
+        alert.setX(scene.getWidth() - 500);
+        alert.setY(500);
+        alert.resultProperty().addListener(r -> {
+            if (alert.getResult().equals(ButtonType.YES)) {
+//                manifoldControlPane.minimize();
+                scene.getRoot().fireEvent(
+                    new ApplicationEvent(ApplicationEvent.SHOW_VIDEO_PANE,
+                        "EMPTY VISION ", "A past never had for a Retrowave Future"));
+            }
+        });
+        alert.show();
+
+        setOnSucceeded(e -> {
+            Platform.runLater(() -> {
+                scene.getRoot().fireEvent(
+                    new ApplicationEvent(ApplicationEvent.HIDE_BUSY_INDICATOR));
+            });
+        });
+        setOnFailed(e -> {
+            Platform.runLater(() -> {
+                scene.getRoot().fireEvent(
+                    new ApplicationEvent(ApplicationEvent.HIDE_BUSY_INDICATOR));
+            });
+        });
+        setOnCancelled(e -> {
+            Platform.runLater(() -> {
+                scene.getRoot().fireEvent(
+                    new ApplicationEvent(ApplicationEvent.HIDE_BUSY_INDICATOR));
+            });
+        });
+   
+    }
+
+    @Override
+    protected FeatureCollection call() throws Exception {
+        if (isCancelled()) return null;
+        Platform.runLater(() -> {
+            ProgressStatus ps = new ProgressStatus("Fitting UMAP Transform...", 0.5);
+            ps.fillStartColor = Color.AZURE;
+            ps.fillEndColor = Color.LIME;
+            ps.innerStrokeColor = Color.AZURE;
+            ps.outerStrokeColor = Color.LIME;
+            scene.getRoot().fireEvent(
+                new ApplicationEvent(ApplicationEvent.UPDATE_BUSY_INDICATOR, ps));
+        });
+
+        double[][] umapMatrix = AnalysisUtils.fitUMAP(originalFC, umap);
+        Platform.runLater(() -> {
+            ProgressStatus ps = new ProgressStatus("Converting to FeatureCollection...", 0.5);
+            ps.fillStartColor = Color.CYAN;
+            ps.fillEndColor = Color.NAVY;
+            ps.innerStrokeColor = Color.CYAN;
+            ps.outerStrokeColor = Color.SILVER;
+            scene.getRoot().fireEvent(
+                new ApplicationEvent(ApplicationEvent.UPDATE_BUSY_INDICATOR, ps));
+        });
+        System.out.println("mapping projected UMAP data back to FeatureVectors...");
+        FeatureCollection projectedFC = FeatureCollection.fromData(umapMatrix);
+        for (int i = 0; i < originalFC.getFeatures().size(); i++) {
+            FeatureVector origFV = originalFC.getFeatures().get(i);
+            projectedFC.getFeatures().get(i).setLabel(origFV.getLabel());
+            projectedFC.getFeatures().get(i).setScore(origFV.getScore());
+            projectedFC.getFeatures().get(i).setImageURL(origFV.getImageURL());
+            projectedFC.getFeatures().get(i).setText(origFV.getText());
+            projectedFC.getFeatures().get(i).setMetaData(origFV.getMetaData());
+        }
+        Platform.runLater(() -> {
+            ProgressStatus ps = new ProgressStatus("", -1);
+            scene.getRoot().fireEvent(
+                new ApplicationEvent(ApplicationEvent.HIDE_BUSY_INDICATOR, ps));
+        });
+        return projectedFC;
+    }
+    
+    /**
+     * @return the cancelledByUser
+     */
+    public boolean isCancelledByUser() {
+        return cancelledByUser;
+    }
+
+    /**
+     * @param cancelledByUser the cancelledByUser to set
+     */
+    public void setCancelledByUser(boolean cancelledByUser) {
+        this.cancelledByUser = cancelledByUser;
+    }
+
+    /**
+     * @return the projectionScalar
+     */
+    public double getProjectionScalar() {
+        return projectionScalar;
+    }
+
+    /**
+     * @param projectionScalar the projectionScalar to set
+     */
+    public void setProjectionScalar(double projectionScalar) {
+        this.projectionScalar = projectionScalar;
+    }
+}
